Create SECURITY.md
Is it true Rainbow is falsely advertising "free mint" on its mobile app, but then charging individuals ETH, (distinct from, and in addition to, the transactional gas fee) without their knowledge or consent?

And then is it true Rainbow provides no means of communicating to resolve this matter except a singular support email which Rainbow ignores, when receiving communications regarding its false advertising?

Is it true Rainbow then gaslights its victims when they use Github to discuss their fraud, and misleadingly informs victims to communicate via the email address that Rainbow fails to adequately monitor?

Is it true that when Rainbow fails to resolve its fraud in a timely manner, and is provided an opportunity to resolve the matter amicably, Rainbow resorts to blocking its victims from Rainbow's Github (the only place where Rainbow responds to its victims)

Does Rainbow realize fraud is a crime in America?

Please ensure these questions are answered. Do not try to gaslight me.